changed how connections are made
