+++ Gemfile
@@ -1,4 +1,2 @@
 source :rubygems
-gem "awesome_print"
-gem "json"
-gem "curb"
+gemspec
